make create idempotent
